revert changes to emscripten
