+++ Cargo.toml
@@ -45,4 +45,4 @@ debug = ["wasmer-clif-backend/debug", "wasmer-runtime-core/debug"]
 fast-tests = []
 llvm = ["wasmer-llvm-backend"]
 dynasm = ["wasmer-dynasm-backend"]
-vfs = ["wasmer-runtime-abi", "wasmer-emscripten/vfs"]
+vfs = ["wasmer-runtime-abi"]
