+++ lib/emscripten/Cargo.toml
@@ -9,19 +9,14 @@ edition = "2018"
 build = "build/mod.rs"
 
 [dependencies]
-bit_field = "0.9.0"
-byteorder = "1"
-errno = "0.2.4"
+wasmer-runtime-core = { path = "../runtime-core", version = "0.2.1" }
 lazy_static = "1.2.0"
 libc = "0.2.49"
+byteorder = "1"
 time = "0.1.41"
-wasmer-runtime-core = { path = "../runtime-core", version = "0.2.1" }
 wasmer-clif-backend = { path = "../clif-backend", version = "0.2.0" }
 wasmer-dynasm-backend = { path = "../dynasm-backend", version = "0.1.0", optional = true }
 wasmer-llvm-backend = { path = "../llvm-backend", version = "0.1.0", optional = true }
-
-[target.'cfg(unix)'.dependencies]
-wasmer-runtime-abi = { path = "../runtime-abi", optional = true }
 
 [target.'cfg(windows)'.dependencies]
 rand = "0.6"
@@ -36,4 +31,3 @@ glob = "0.2.11"
 clif = []
 llvm = ["wasmer-llvm-backend"]
 dynasm = ["wasmer-dynasm-backend"]
-vfs = ["wasmer-runtime-abi"]
