+++ lib/emscripten/src/syscalls/mod.rs
@@ -4,26 +4,44 @@ mod unix;
 #[cfg(windows)]
 mod windows;
 
-#[cfg(all(feature = "vfs", not(target_os = "windows")))]
-mod emscripten_vfs;
-
 #[cfg(unix)]
 pub use self::unix::*;
 
 #[cfg(windows)]
 pub use self::windows::*;
 
-#[cfg(all(feature = "vfs", not(target_os = "windows")))]
-pub use self::emscripten_vfs::*;
-
+use super::utils::copy_stat_into_wasm;
 use super::varargs::VarArgs;
 use byteorder::{ByteOrder, LittleEndian};
 /// NOTE: TODO: These syscalls only support wasm_32 for now because they assume offsets are u32
 /// Syscall list: https://www.cs.utexas.edu/~bismith/test/syscalls/syscalls32.html
-use libc::{c_int, c_void, chdir, exit, getpid, lseek, rmdir};
+use libc::{
+    // ENOTTY,
+    c_int,
+    c_void,
+    chdir,
+    // fcntl, setsockopt, getppid
+    close,
+    dup2,
+    exit,
+    fstat,
+    getpid,
+    // iovec,
+    lseek,
+    off_t,
+    //    open,
+    read,
+    // readv,
+    rmdir,
+    // writev,
+    stat,
+    write,
+    // sockaddr_in,
+};
 use wasmer_runtime_core::vm::Ctx;
 
 use super::env;
+use std::cell::Cell;
 #[allow(unused_imports)]
 use std::io::Error;
 use std::mem;
@@ -36,6 +54,39 @@ pub fn ___syscall1(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) {
     unsafe {
         exit(status);
     }
+}
+
+/// read
+pub fn ___syscall3(ctx: &mut Ctx, _which: i32, mut varargs: VarArgs) -> i32 {
+    // -> ssize_t
+    debug!("emscripten::___syscall3 (read) {}", _which);
+    let fd: i32 = varargs.get(ctx);
+    let buf: u32 = varargs.get(ctx);
+    let count: i32 = varargs.get(ctx);
+    debug!("=> fd: {}, buf_offset: {}, count: {}", fd, buf, count);
+    let buf_addr = emscripten_memory_pointer!(ctx.memory(0), buf) as *mut c_void;
+    let ret = unsafe { read(fd, buf_addr, count as _) };
+    debug!("=> ret: {}", ret);
+    ret as _
+}
+
+/// write
+pub fn ___syscall4(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
+    debug!("emscripten::___syscall4 (write) {}", _which);
+    let fd: i32 = varargs.get(ctx);
+    let buf: u32 = varargs.get(ctx);
+    let count: i32 = varargs.get(ctx);
+    debug!("=> fd: {}, buf: {}, count: {}", fd, buf, count);
+    let buf_addr = emscripten_memory_pointer!(ctx.memory(0), buf) as *const c_void;
+    unsafe { write(fd, buf_addr, count as _) as i32 }
+}
+
+/// close
+pub fn ___syscall6(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
+    debug!("emscripten::___syscall6 (close) {}", _which);
+    let fd: i32 = varargs.get(ctx);
+    debug!("fd: {}", fd);
+    unsafe { close(fd) }
 }
 
 // chdir
@@ -53,6 +104,11 @@ pub fn ___syscall12(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int
 
 pub fn ___syscall10(_ctx: &mut Ctx, _one: i32, _two: i32) -> i32 {
     debug!("emscripten::___syscall10");
+    -1
+}
+
+pub fn ___syscall15(_ctx: &mut Ctx, _one: i32, _two: i32) -> i32 {
+    debug!("emscripten::___syscall15");
     -1
 }
 
@@ -75,16 +131,51 @@ pub fn ___syscall40(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int
     unsafe { rmdir(pathname_addr) }
 }
 
+// pipe
+pub fn ___syscall42(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
+    debug!("emscripten::___syscall42 (pipe)");
+    // offset to a file descriptor, which contains a read end and write end, 2 integers
+    let fd_offset: u32 = varargs.get(ctx);
+
+    let emscripten_memory = ctx.memory(0);
+
+    // convert the file descriptor into a vec with two slots
+    let mut fd_vec: Vec<c_int> = emscripten_memory.view()[((fd_offset / 4) as usize)..]
+        .iter()
+        .map(|pipe_end: &Cell<c_int>| pipe_end.get())
+        .take(2)
+        .collect();
+
+    // get it as a mutable pointer
+    let fd_ptr = fd_vec.as_mut_ptr();
+
+    // call pipe and store the pointers in this array
+    #[cfg(target_os = "windows")]
+    let result: c_int = unsafe { libc::pipe(fd_ptr, 2048, 0) };
+    #[cfg(not(target_os = "windows"))]
+    let result: c_int = unsafe { libc::pipe(fd_ptr) };
+    result
+}
+
 pub fn ___syscall60(_ctx: &mut Ctx, _one: i32, _two: i32) -> i32 {
     debug!("emscripten::___syscall60");
     -1
 }
 
+// dup2
+pub fn ___syscall63(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
+    debug!("emscripten::___syscall63 (dup2) {}", _which);
+
+    let src: i32 = varargs.get(ctx);
+    let dst: i32 = varargs.get(ctx);
+
+    unsafe { dup2(src, dst) }
+}
+
 // getppid
 pub fn ___syscall64(_ctx: &mut Ctx, _one: i32, _two: i32) -> i32 {
     debug!("emscripten::___syscall64 (getppid)");
-    let result = unsafe { getpid() };
-    result
+    unsafe { getpid() }
 }
 
 pub fn ___syscall66(_ctx: &mut Ctx, _one: i32, _two: i32) -> i32 {
@@ -173,7 +264,7 @@ pub fn ___syscall140(ctx: &mut Ctx, _which: i32, mut varargs: VarArgs) -> i32 {
     let offset_low: i32 = varargs.get(ctx);
     let result_ptr_value = varargs.get::<i32>(ctx);
     let whence: i32 = varargs.get(ctx);
-    let offset = offset_low as libc::off_t;
+    let offset = offset_low as off_t;
     let ret = unsafe { lseek(fd, offset, whence) as i32 };
     #[allow(clippy::cast_ptr_alignment)]
     let result_ptr = emscripten_memory_pointer!(ctx.memory(0), result_ptr_value) as *mut i32;
@@ -219,7 +310,43 @@ pub fn ___syscall145(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> i32 
                 as *mut c_void;
             let iov_len = (*guest_iov_addr).iov_len as _;
             // debug!("=> iov_addr: {:?}, {:?}", iov_base, iov_len);
-            let curr = libc::read(fd, iov_base, iov_len);
+            let curr = read(fd, iov_base, iov_len);
+            if curr < 0 {
+                return -1;
+            }
+            ret += curr;
+        }
+        // debug!(" => ret: {}", ret);
+        ret as _
+    }
+}
+
+// writev
+#[allow(clippy::cast_ptr_alignment)]
+pub fn ___syscall146(ctx: &mut Ctx, _which: i32, mut varargs: VarArgs) -> i32 {
+    // -> ssize_t
+    debug!("emscripten::___syscall146 (writev) {}", _which);
+    let fd: i32 = varargs.get(ctx);
+    let iov: i32 = varargs.get(ctx);
+    let iovcnt: i32 = varargs.get(ctx);
+
+    #[repr(C)]
+    struct GuestIovec {
+        iov_base: i32,
+        iov_len: i32,
+    }
+
+    debug!("=> fd: {}, iov: {}, iovcnt = {}", fd, iov, iovcnt);
+    let mut ret = 0;
+    unsafe {
+        for i in 0..iovcnt {
+            let guest_iov_addr =
+                emscripten_memory_pointer!(ctx.memory(0), (iov + i * 8)) as *mut GuestIovec;
+            let iov_base = emscripten_memory_pointer!(ctx.memory(0), (*guest_iov_addr).iov_base)
+                as *const c_void;
+            let iov_len = (*guest_iov_addr).iov_len as _;
+            // debug!("=> iov_addr: {:?}, {:?}", iov_base, iov_len);
+            let curr = write(fd, iov_base, iov_len);
             if curr < 0 {
                 return -1;
             }
@@ -254,8 +381,8 @@ pub fn ___syscall195(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_in
     let pathname_addr = emscripten_memory_pointer!(ctx.memory(0), pathname) as *const i8;
 
     unsafe {
-        let mut _stat: libc::stat = std::mem::zeroed();
-        let ret = libc::stat(pathname_addr, &mut _stat);
+        let mut _stat: stat = std::mem::zeroed();
+        let ret = stat(pathname_addr, &mut _stat);
         debug!(
             "=> pathname: {}, buf: {}, path: {} = {}\nlast os error: {}",
             pathname,
@@ -267,7 +394,7 @@ pub fn ___syscall195(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_in
         if ret != 0 {
             return ret;
         }
-        crate::utils::copy_stat_into_wasm(ctx, buf, &_stat);
+        copy_stat_into_wasm(ctx, buf, &_stat);
     }
     0
 }
@@ -280,12 +407,12 @@ pub fn ___syscall197(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_in
 
     unsafe {
         let mut stat = std::mem::zeroed();
-        let ret = libc::fstat(fd, &mut stat);
+        let ret = fstat(fd, &mut stat);
         debug!("ret: {}", ret);
         if ret != 0 {
             return ret;
         }
-        crate::utils::copy_stat_into_wasm(ctx, buf, &stat);
+        copy_stat_into_wasm(ctx, buf, &stat);
     }
 
     0
