--- lib/emscripten/src/syscalls/unix/host_fs.rs
@@ -1,575 +1,0 @@
-use crate::utils::copy_stat_into_wasm;
-use crate::varargs::VarArgs;
-use libc::{c_int, c_void, ioctl, sockaddr, socklen_t};
-use wasmer_runtime_core::vm::Ctx;
-
-/// read
-pub fn ___syscall3(ctx: &mut Ctx, _which: i32, mut varargs: VarArgs) -> i32 {
-    debug!("emscripten::___syscall3 (read) {}", _which);
-    let fd: i32 = varargs.get(ctx);
-    let buf: u32 = varargs.get(ctx);
-    let count: i32 = varargs.get(ctx);
-    debug!("=> fd: {}, buf_offset: {}, count: {}", fd, buf, count);
-    let buf_addr = emscripten_memory_pointer!(ctx.memory(0), buf) as *mut c_void;
-    let ret = unsafe { libc::read(fd, buf_addr, count as _) };
-    debug!("=> ret: {}", ret);
-    ret as _
-}
-
-/// write
-pub fn ___syscall4(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
-    debug!("emscripten::___syscall4 (write) {}", _which);
-    let fd: i32 = varargs.get(ctx);
-    let buf: u32 = varargs.get(ctx);
-    let count: i32 = varargs.get(ctx);
-    debug!("=> fd: {}, buf: {}, count: {}", fd, buf, count);
-    let buf_addr = emscripten_memory_pointer!(ctx.memory(0), buf) as *const c_void;
-    let ret = unsafe { libc::write(fd, buf_addr, count as _) as i32 };
-    ret
-}
-
-/// open
-pub fn ___syscall5(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
-    debug!("emscripten::___syscall5 (open) {}", _which);
-    let pathname: u32 = varargs.get(ctx);
-    let flags: i32 = varargs.get(ctx);
-    let mode: u32 = varargs.get(ctx);
-    let pathname_addr = emscripten_memory_pointer!(ctx.memory(0), pathname) as *const i8;
-    let _path_str = unsafe { std::ffi::CStr::from_ptr(pathname_addr).to_str().unwrap() };
-    let fd = unsafe { libc::open(pathname_addr, flags, mode) };
-    debug!(
-        "=> pathname: {}, flags: {}, mode: {} = fd: {}\npath: {}",
-        pathname, flags, mode, fd, _path_str
-    );
-    fd
-}
-
-/// close
-pub fn ___syscall6(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
-    debug!("emscripten::___syscall6 (close) {}", _which);
-    let fd: i32 = varargs.get(ctx);
-    debug!("fd: {}", fd);
-    unsafe { libc::close(fd) }
-}
-
-/// chmod
-pub fn ___syscall15(_ctx: &mut Ctx, _one: i32, _two: i32) -> i32 {
-    debug!("emscripten::___syscall15");
-    -1
-}
-
-/// mkdir
-pub fn ___syscall39(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
-    debug!("emscripten::___syscall39 (mkdir) {}", _which);
-    let pathname: u32 = varargs.get(ctx);
-    let mode: u32 = varargs.get(ctx);
-    let pathname_addr = emscripten_memory_pointer!(ctx.memory(0), pathname) as *const i8;
-    unsafe { libc::mkdir(pathname_addr, mode as _) }
-}
-
-/// pipe
-pub fn ___syscall42(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
-    debug!("emscripten::___syscall42 (pipe)");
-    // offset to a file descriptor, which contains a read end and write end, 2 integers
-    let fd_offset: u32 = varargs.get(ctx);
-
-    let emscripten_memory = ctx.memory(0);
-
-    // convert the file descriptor into a vec with two slots
-    let mut fd_vec: Vec<c_int> = emscripten_memory.view()[((fd_offset / 4) as usize)..]
-        .iter()
-        .map(|pipe_end: &std::cell::Cell<c_int>| pipe_end.get())
-        .take(2)
-        .collect();
-
-    // get it as a mutable pointer
-    let fd_ptr = fd_vec.as_mut_ptr();
-
-    // call pipe and store the pointers in this array
-    #[cfg(target_os = "windows")]
-    let result: c_int = unsafe { libc::pipe(fd_ptr, 2048, 0) };
-    #[cfg(not(target_os = "windows"))]
-    let result: c_int = unsafe { libc::pipe(fd_ptr) };
-    result
-}
-
-/// ioctl
-pub fn ___syscall54(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
-    debug!("emscripten::___syscall54 (ioctl) {}", _which);
-    let fd: i32 = varargs.get(ctx);
-    let request: u32 = varargs.get(ctx);
-    debug!("fd: {}, op: {}", fd, request);
-    // Got the equivalents here: https://code.woboq.org/linux/linux/include/uapi/asm-generic/ioctls.h.html
-    match request as _ {
-        21537 => {
-            // FIONBIO
-            let argp: u32 = varargs.get(ctx);
-            let argp_ptr = emscripten_memory_pointer!(ctx.memory(0), argp) as *mut c_void;
-            let ret = unsafe { ioctl(fd, libc::FIONBIO, argp_ptr) };
-            debug!("ret(FIONBIO): {}", ret);
-            ret
-            // 0
-        }
-        21523 => {
-            // TIOCGWINSZ
-            let argp: u32 = varargs.get(ctx);
-            let argp_ptr = emscripten_memory_pointer!(ctx.memory(0), argp) as *mut c_void;
-            let ret = unsafe { ioctl(fd, libc::TIOCGWINSZ, argp_ptr) };
-            debug!("ret(TIOCGWINSZ): {} (harcoded to 0)", ret);
-            // ret
-            // TODO: We hardcode the value to have emscripten tests pass, as for some reason
-            // when the capturer is active, ioctl returns -1 instead of 0
-            if ret == -1 {
-                0
-            } else {
-                ret
-            }
-        }
-        _ => {
-            debug!(
-                "emscripten::___syscall54 -> non implemented case {}",
-                request
-            );
-            0
-        }
-    }
-}
-
-/// dup2
-pub fn ___syscall63(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
-    debug!("emscripten::___syscall63 (dup2) {}", _which);
-
-    let src: i32 = varargs.get(ctx);
-    let dst: i32 = varargs.get(ctx);
-
-    unsafe { libc::dup2(src, dst) }
-}
-
-// socketcall
-#[allow(clippy::cast_ptr_alignment)]
-pub fn ___syscall102(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
-    debug!("emscripten::___syscall102 (socketcall) {}", _which);
-    let call: u32 = varargs.get(ctx);
-    let mut socket_varargs: VarArgs = varargs.get(ctx);
-
-    #[repr(C)]
-    pub struct GuestSockaddrIn {
-        pub sin_family: libc::sa_family_t,
-        // u16
-        pub sin_port: libc::in_port_t,
-        // u16
-        pub sin_addr: GuestInAddr,
-        // u32
-        pub sin_zero: [u8; 8], // u8 * 8
-                               // 2 + 2 + 4 + 8 = 16
-    }
-
-    #[repr(C)]
-    pub struct GuestInAddr {
-        pub s_addr: libc::in_addr_t, // u32
-    }
-
-    // debug!("GuestSockaddrIn = {}", size_of::<GuestSockaddrIn>());
-
-    pub struct LinuxSockAddr {
-        pub sa_family: u16,
-        pub sa_data: [libc::c_char; 14],
-    }
-
-    match call {
-        1 => {
-            debug!("socket: socket");
-            // Another conditional constant for name resolution: Macos et iOS use
-            // SO_NOSIGPIPE as a setsockopt flag to disable SIGPIPE emission on socket.
-            // Other platforms do otherwise.
-            #[cfg(target_os = "darwin")]
-            use libc::SO_NOSIGPIPE;
-            #[cfg(not(target_os = "darwin"))]
-            const SO_NOSIGPIPE: c_int = 0;
-
-            // socket (domain: c_int, ty: c_int, protocol: c_int) -> c_int
-            let domain: i32 = socket_varargs.get(ctx);
-            let ty: i32 = socket_varargs.get(ctx);
-            let protocol: i32 = socket_varargs.get(ctx);
-            let fd = unsafe { libc::socket(domain, ty, protocol) };
-            // set_cloexec
-            unsafe {
-                ioctl(fd, libc::FIOCLEX);
-            };
-
-            let _err = errno::errno();
-
-            let _result =
-                unsafe { libc::setsockopt(fd, libc::SOL_SOCKET, SO_NOSIGPIPE, 0 as *const _, 4) };
-
-            let _err2 = errno::errno();
-
-            debug!(
-                "=> domain: {} (AF_INET/2), type: {} (SOCK_STREAM/1), protocol: {} = fd: {}",
-                domain, ty, protocol, fd
-            );
-            fd as _
-        }
-        2 => {
-            debug!("socket: bind");
-            // bind (socket: c_int, address: *const sockaddr, address_len: socklen_t) -> c_int
-            // TODO: Emscripten has a different signature.
-            let socket = socket_varargs.get(ctx);
-            let address: u32 = socket_varargs.get(ctx);
-            let address_len = socket_varargs.get(ctx);
-            let address = emscripten_memory_pointer!(ctx.memory(0), address) as *mut sockaddr;
-
-            // Debug received address
-            let _proper_address = address as *const GuestSockaddrIn;
-            debug!(
-                "=> address.sin_family: {:?}, address.sin_port: {:?}, address.sin_addr.s_addr: {:?}",
-                unsafe { (*_proper_address).sin_family }, unsafe { (*_proper_address).sin_port }, unsafe { (*_proper_address).sin_addr.s_addr }
-            );
-            let status = unsafe { libc::bind(socket, address, address_len) };
-            // debug!("=> status: {}", status);
-            debug!(
-                "=> socketfd: {}, address: {:?}, address_len: {} = status: {}",
-                socket, address, address_len, status
-            );
-            status
-            // -1
-        }
-        3 => {
-            debug!("socket: connect");
-            // connect (socket: c_int, address: *const sockaddr, len: socklen_t) -> c_int
-            // TODO: Emscripten has a different signature.
-            let socket = socket_varargs.get(ctx);
-            let address: u32 = socket_varargs.get(ctx);
-            let address_len = socket_varargs.get(ctx);
-            let address = emscripten_memory_pointer!(ctx.memory(0), address) as *mut sockaddr;
-            unsafe { libc::connect(socket, address, address_len) }
-        }
-        4 => {
-            debug!("socket: listen");
-            // listen (socket: c_int, backlog: c_int) -> c_int
-            let socket = socket_varargs.get(ctx);
-            let backlog: i32 = socket_varargs.get(ctx);
-            let status = unsafe { libc::listen(socket, backlog) };
-            debug!(
-                "=> socketfd: {}, backlog: {} = status: {}",
-                socket, backlog, status
-            );
-            status
-        }
-        5 => {
-            debug!("socket: accept");
-            // accept (socket: c_int, address: *mut sockaddr, address_len: *mut socklen_t) -> c_int
-            let socket = socket_varargs.get(ctx);
-            let address_addr: u32 = socket_varargs.get(ctx);
-            let address_len: u32 = socket_varargs.get(ctx);
-            let address = emscripten_memory_pointer!(ctx.memory(0), address_addr) as *mut sockaddr;
-
-            debug!(
-                "=> socket: {}, address: {:?}, address_len: {}",
-                socket, address, address_len
-            );
-            let address_len_addr =
-                emscripten_memory_pointer!(ctx.memory(0), address_len) as *mut socklen_t;
-            // let mut address_len_addr: socklen_t = 0;
-
-            let fd = unsafe { libc::accept(socket, address, address_len_addr) };
-
-            unsafe {
-                let address_linux =
-                    emscripten_memory_pointer!(ctx.memory(0), address_addr) as *mut LinuxSockAddr;
-                (*address_linux).sa_family = (*address).sa_family as u16;
-                (*address_linux).sa_data = (*address).sa_data;
-                let _proper_address = address as *const GuestSockaddrIn;
-                let _x = 10;
-            };
-
-            // set_cloexec
-            unsafe {
-                ioctl(fd, libc::FIOCLEX);
-            };
-
-            debug!("fd: {}", fd);
-
-            fd as _
-        }
-        6 => {
-            debug!("socket: getsockname");
-            // getsockname (socket: c_int, address: *mut sockaddr, address_len: *mut socklen_t) -> c_int
-            let socket = socket_varargs.get(ctx);
-            let address: u32 = socket_varargs.get(ctx);
-            let address_len: u32 = socket_varargs.get(ctx);
-            let address = emscripten_memory_pointer!(ctx.memory(0), address) as *mut sockaddr;
-            let address_len_addr =
-                emscripten_memory_pointer!(ctx.memory(0), address_len) as *mut socklen_t;
-            unsafe { libc::getsockname(socket, address, address_len_addr) }
-        }
-        7 => {
-            debug!("socket: getpeername");
-            // getpeername (socket: c_int, address: *mut sockaddr, address_len: *mut socklen_t) -> c_int
-            let socket = socket_varargs.get(ctx);
-            let address: u32 = socket_varargs.get(ctx);
-            let address_len: u32 = socket_varargs.get(ctx);
-            let address = emscripten_memory_pointer!(ctx.memory(0), address) as *mut sockaddr;
-            let address_len_addr =
-                emscripten_memory_pointer!(ctx.memory(0), address_len) as *mut socklen_t;
-            unsafe { libc::getpeername(socket, address, address_len_addr) }
-        }
-        11 => {
-            debug!("socket: sendto");
-            // sendto (socket: c_int, buf: *const c_void, len: size_t, flags: c_int, addr: *const sockaddr, addrlen: socklen_t) -> ssize_t
-            let socket = socket_varargs.get(ctx);
-            let buf: u32 = socket_varargs.get(ctx);
-            let flags = socket_varargs.get(ctx);
-            let len: i32 = socket_varargs.get(ctx);
-            let address: u32 = socket_varargs.get(ctx);
-            let address_len = socket_varargs.get(ctx);
-            let buf_addr = emscripten_memory_pointer!(ctx.memory(0), buf) as _;
-            let address = emscripten_memory_pointer!(ctx.memory(0), address) as *mut sockaddr;
-            unsafe { libc::sendto(socket, buf_addr, flags, len, address, address_len) as i32 }
-        }
-        12 => {
-            debug!("socket: recvfrom");
-            // recvfrom (socket: c_int, buf: *const c_void, len: size_t, flags: c_int, addr: *const sockaddr, addrlen: socklen_t) -> ssize_t
-            let socket: i32 = socket_varargs.get(ctx);
-            let buf: u32 = socket_varargs.get(ctx);
-            let flags = socket_varargs.get(ctx);
-            let len: i32 = socket_varargs.get(ctx);
-            let address: u32 = socket_varargs.get(ctx);
-            let address_len: u32 = socket_varargs.get(ctx);
-            let buf_addr = emscripten_memory_pointer!(ctx.memory(0), buf) as _;
-            let address = emscripten_memory_pointer!(ctx.memory(0), address) as *mut sockaddr;
-            let address_len_addr =
-                emscripten_memory_pointer!(ctx.memory(0), address_len) as *mut socklen_t;
-            let recv_result = unsafe {
-                libc::recvfrom(socket, buf_addr, flags, len, address, address_len_addr) as i32
-            };
-            debug!(
-                "recvfrom: socket: {}, flags: {}, len: {}, result: {}",
-                socket, flags, len, recv_result
-            );
-            recv_result
-        }
-        14 => {
-            debug!("socket: setsockopt");
-            // NOTE: Emscripten seems to be passing the wrong values to this syscall
-            //      level: Em passes 1 as SOL_SOCKET; SOL_SOCKET is 0xffff in BSD
-            //      name: Em passes SO_ACCEPTCONN, but Nginx complains about REUSEADDR
-            //      https://github.com/openbsd/src/blob/master/sys/sys/socket.h#L156
-            // setsockopt (socket: c_int, level: c_int, name: c_int, value: *const c_void, option_len: socklen_t) -> c_int
-
-            let socket = socket_varargs.get(ctx);
-            // SOL_SOCKET = 0xffff (BSD, Linux)
-            let level: i32 = libc::SOL_SOCKET;
-            let _: u32 = socket_varargs.get(ctx);
-            // SO_REUSEADDR = 0x4 (BSD, Linux)
-            let name: i32 = libc::SO_REUSEADDR;
-            let _: u32 = socket_varargs.get(ctx);
-            let value: u32 = socket_varargs.get(ctx);
-            let option_len = socket_varargs.get(ctx);
-            let value_addr = emscripten_memory_pointer!(ctx.memory(0), value) as _; // Endian problem
-            let ret = unsafe { libc::setsockopt(socket, level, name, value_addr, option_len) };
-
-            debug!("=> socketfd: {}, level: {} (SOL_SOCKET/0xffff), name: {} (SO_REUSEADDR/4), value_addr: {:?}, option_len: {} = status: {}", socket, level, name, value_addr, option_len, ret);
-            ret
-        }
-        15 => {
-            debug!("socket: getsockopt");
-            // getsockopt (sockfd: c_int, level: c_int, optname: c_int, optval: *mut c_void, optlen: *mut socklen_t) -> c_int
-            let socket = socket_varargs.get(ctx);
-            let level: i32 = socket_varargs.get(ctx);
-            let name: i32 = socket_varargs.get(ctx);
-            let value: u32 = socket_varargs.get(ctx);
-            let option_len: u32 = socket_varargs.get(ctx);
-            let value_addr = emscripten_memory_pointer!(ctx.memory(0), value) as _;
-            let option_len_addr =
-                emscripten_memory_pointer!(ctx.memory(0), option_len) as *mut socklen_t;
-            let result =
-                unsafe { libc::getsockopt(socket, level, name, value_addr, option_len_addr) };
-            result
-        }
-        16 => {
-            debug!("socket: sendmsg");
-            // sendmsg (fd: c_int, msg: *const msghdr, flags: c_int) -> ssize_t
-            let socket: i32 = socket_varargs.get(ctx);
-            let msg: u32 = socket_varargs.get(ctx);
-            let flags: i32 = socket_varargs.get(ctx);
-            let msg_addr = emscripten_memory_pointer!(ctx.memory(0), msg) as *const libc::msghdr;
-            unsafe { libc::sendmsg(socket, msg_addr, flags) as i32 }
-        }
-        17 => {
-            debug!("socket: recvmsg");
-            // recvmsg (fd: c_int, msg: *mut msghdr, flags: c_int) -> ssize_t
-            let socket: i32 = socket_varargs.get(ctx);
-            let msg: u32 = socket_varargs.get(ctx);
-            let flags: i32 = socket_varargs.get(ctx);
-            let msg_addr = emscripten_memory_pointer!(ctx.memory(0), msg) as *mut libc::msghdr;
-            unsafe { libc::recvmsg(socket, msg_addr, flags) as i32 }
-        }
-        _ => {
-            // others
-            -1
-        }
-    }
-}
-
-/// select = ___syscall142
-#[allow(clippy::cast_ptr_alignment)]
-pub fn ___syscall142(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
-    debug!("emscripten::___syscall142 (newselect) {}", _which);
-
-    let nfds: i32 = varargs.get(ctx);
-    let readfds: u32 = varargs.get(ctx);
-    let writefds: u32 = varargs.get(ctx);
-    let _exceptfds: u32 = varargs.get(ctx);
-    let _timeout: i32 = varargs.get(ctx);
-    let readfds_ptr = emscripten_memory_pointer!(ctx.memory(0), readfds) as _;
-    let writefds_ptr = emscripten_memory_pointer!(ctx.memory(0), writefds) as _;
-    let result = unsafe { libc::select(nfds, readfds_ptr, writefds_ptr, 0 as _, 0 as _) };
-    debug!("select returns {}", result);
-    result
-}
-
-// writev
-#[allow(clippy::cast_ptr_alignment)]
-pub fn ___syscall146(ctx: &mut Ctx, _which: i32, mut varargs: VarArgs) -> i32 {
-    // -> ssize_t
-    debug!("emscripten::___syscall146 (writev) {}", _which);
-    let fd: i32 = varargs.get(ctx);
-    let iov: i32 = varargs.get(ctx);
-    let iovcnt: i32 = varargs.get(ctx);
-
-    #[repr(C)]
-    struct GuestIovec {
-        iov_base: i32,
-        iov_len: i32,
-    }
-
-    debug!("=> fd: {}, iov: {}, iovcnt = {}", fd, iov, iovcnt);
-    let mut ret = 0;
-    unsafe {
-        for i in 0..iovcnt {
-            let guest_iov_addr =
-                emscripten_memory_pointer!(ctx.memory(0), (iov + i * 8)) as *mut GuestIovec;
-            let iov_base = emscripten_memory_pointer!(ctx.memory(0), (*guest_iov_addr).iov_base)
-                as *const c_void;
-            let iov_len = (*guest_iov_addr).iov_len as _;
-            // debug!("=> iov_addr: {:?}, {:?}", iov_base, iov_len);
-            let curr = libc::write(fd, iov_base, iov_len);
-            if curr < 0 {
-                return -1;
-            }
-            ret += curr;
-        }
-        // debug!(" => ret: {}", ret);
-        ret as _
-    }
-}
-
-/// pread
-pub fn ___syscall180(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
-    debug!("emscripten::___syscall180 (pread) {}", _which);
-    let fd: i32 = varargs.get(ctx);
-    let buf: u32 = varargs.get(ctx);
-    let count: u32 = varargs.get(ctx);
-    {
-        let zero: u32 = varargs.get(ctx);
-        assert_eq!(zero, 0);
-    }
-    let offset: i64 = varargs.get(ctx);
-    let buf_ptr = emscripten_memory_pointer!(ctx.memory(0), buf) as _;
-    let pread_result = unsafe { libc::pread(fd, buf_ptr, count as _, offset) as _ };
-    pread_result
-}
-
-/// pwrite
-pub fn ___syscall181(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
-    debug!("emscripten::___syscall181 (pwrite) {}", _which);
-    let fd: i32 = varargs.get(ctx);
-    let buf: u32 = varargs.get(ctx);
-    let count: u32 = varargs.get(ctx);
-    {
-        let zero: u32 = varargs.get(ctx);
-        assert_eq!(zero, 0);
-    }
-    let offset: i64 = varargs.get(ctx);
-    let buf_ptr = emscripten_memory_pointer!(ctx.memory(0), buf) as _;
-    let status = unsafe { libc::pwrite(fd, buf_ptr, count as _, offset) as _ };
-    debug!(
-        "=> fd: {}, buf: {}, count: {}, offset: {} = status:{}",
-        fd, buf, count, offset, status
-    );
-    status
-}
-
-/// stat64
-pub fn ___syscall195(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
-    debug!("emscripten::___syscall195 (stat64) {}", _which);
-    let pathname: u32 = varargs.get(ctx);
-    let buf: u32 = varargs.get(ctx);
-
-    let pathname_addr = emscripten_memory_pointer!(ctx.memory(0), pathname) as *const i8;
-
-    unsafe {
-        let mut _stat: libc::stat = std::mem::zeroed();
-        let ret = libc::stat(pathname_addr, &mut _stat);
-        debug!("ret: {}", ret);
-        if ret != 0 {
-            return ret;
-        }
-        copy_stat_into_wasm(ctx, buf, &_stat);
-    }
-    0
-}
-
-/// fstat64
-pub fn ___syscall197(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> c_int {
-    debug!("emscripten::___syscall197 (fstat64) {}", _which);
-    let fd: c_int = varargs.get(ctx);
-    let buf: u32 = varargs.get(ctx);
-
-    unsafe {
-        let mut stat = std::mem::zeroed();
-        let ret = libc::fstat(fd, &mut stat);
-        debug!("ret: {}", ret);
-        if ret != 0 {
-            return ret;
-        }
-        copy_stat_into_wasm(ctx, buf, &stat);
-    }
-
-    0
-}
-
-/// dup3
-pub fn ___syscall330(ctx: &mut Ctx, _which: c_int, mut varargs: VarArgs) -> libc::pid_t {
-    // Implementation based on description at https://linux.die.net/man/2/dup3
-    debug!("emscripten::___syscall330 (dup3)");
-    let oldfd: c_int = varargs.get(ctx);
-    let newfd: c_int = varargs.get(ctx);
-    let flags: c_int = varargs.get(ctx);
-
-    if oldfd == newfd {
-        return libc::EINVAL;
-    }
-
-    let res = unsafe { libc::dup2(oldfd, newfd) };
-
-    // Set flags on newfd (https://www.gnu.org/software/libc/manual/html_node/Descriptor-Flags.html)
-    let mut old_flags = unsafe { libc::fcntl(newfd, libc::F_GETFD, 0) };
-
-    if old_flags > 0 {
-        old_flags |= flags;
-    } else if old_flags == 0 {
-        old_flags &= !flags;
-    }
-
-    unsafe {
-        libc::fcntl(newfd, libc::F_SETFD, old_flags);
-    }
-
-    debug!(
-        "=> oldfd: {}, newfd: {}, flags: {} = pid: {}",
-        oldfd, newfd, flags, res
-    );
-    res
-}
